% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/domains.R
\name{supported_domains}
\alias{supported_domains}
\title{Domains this package can generate today}
\usage{
supported_domains()
}
\value{
Character vector of supported domain codes.
}
\description{
Domains this package can generate today
}
